% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{connect_to_l500_dbi}
\alias{connect_to_l500_dbi}
\alias{connect_to_sk_dbi}
\alias{connect_to_kr_dbi}
\title{Connect to an instrument database}
\usage{
connect_to_l500_dbi(db_file_path = "D:/Analyzer.mdb")

connect_to_sk_dbi(user = "fguerrero")

connect_to_kr_dbi(db_file_path = "D:/OppLocal.mdb")
}
\arguments{
\item{db_file_path}{path to database file (.mdb)}

\item{user}{username (for now, "fguerrero")}
}
\value{
An S4 object that inherits from DBI::DBIConnection-class.
}
\description{
Connect to an instrument database
}
