% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_clean_methods}
\alias{h_clean_methods}
\title{Standardize method names from multiple instruments}
\usage{
h_clean_methods(x)
}
\arguments{
\item{x}{WIP return value from wrangle_results()? also works with wrangled_cals}
}
\value{
A character vector without trailing numbers
}
\description{
Standardize method names from multiple instruments
}
