% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lida_500.R
\name{h_l500_format_curve}
\alias{h_l500_format_curve}
\title{Format Lida 500 reaction curves}
\usage{
h_l500_format_curve(ABS_Prim, ABS_Sec)
}
\arguments{
\item{ABS_Prim}{vector of primary wavelength absorbance}

\item{ABS_Sec}{vector of secondary wavelength absorbance}
}
\value{
A tibble with columns "cycle", "abs1", "abs2" and "abs_dif"
}
\description{
Format Lida 500 reaction curves
}
