% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_rm_trailing_numbers}
\alias{h_rm_trailing_numbers}
\title{Remove trailing numbers from a character vector}
\usage{
h_rm_trailing_numbers(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A character vector without trailing numbers
}
\description{
Remove trailing numbers from a character vector
}
