% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sekisui.R
\name{h_sk_item_name_to_no}
\alias{h_sk_item_name_to_no}
\title{Convert ITEM_NAME to ITEM_NO for Sekisui SK500 database
`r lifecycle::badge('experimental')`}
\usage{
h_sk_item_name_to_no(item_names, conn = connect_to_sk_dbi())
}
\arguments{
\item{item_names}{Describir...}

\item{conn}{a connection object like the one returned from connect_to_sk_dbi}
}
\value{
an integer vector of ITEM_NO
}
\description{
Convert ITEM_NAME to ITEM_NO for Sekisui SK500 database
`r lifecycle::badge('experimental')`
}
