% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sekisui.R
\name{h_sk_join_rc_res}
\alias{h_sk_join_rc_res}
\title{Join the RC imported data from Sekisui to sk_results and generates an id_rc
`r lifecycle::badge('experimental')`}
\usage{
h_sk_join_rc_res(x, sk_res)
}
\arguments{
\item{x}{RC imported data from Sekisui.}

\item{sk_res}{object of class "sk_results".}
}
\value{
x with a new column of generated "id_rc"
}
\description{
Join the RC imported data from Sekisui to sk_results and generates an id_rc
`r lifecycle::badge('experimental')`
}
