% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_csv.R
\name{imp_csv_kr_cal}
\alias{imp_csv_kr_cal}
\title{Import Kroma or Kroma Plus calibrations from .csv exported from instrument}
\usage{
imp_csv_kr_cal(x)
}
\arguments{
\item{x}{Path to the .csv file.}
}
\value{
A tibble.
}
\description{
Import Kroma or Kroma Plus calibrations from .csv exported from instrument
}
\examples{
x <- system.file("extdata", "kroma_cal.csv", package = "importr")
import_kroma_cal(x)
}
