% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_csv.R
\name{imp_csv_kr_res}
\alias{imp_csv_kr_res}
\title{Import Kroma or Kroma Plus results from .csv exported from instrument}
\usage{
imp_csv_kr_res(x)
}
\arguments{
\item{x}{Path to the .csv file.}
}
\value{
A tibble.
}
\description{
Import Kroma or Kroma Plus results from .csv exported from instrument
}
