% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_csv.R
\name{imp_csv_sk_rc}
\alias{imp_csv_sk_rc}
\title{Import Sekisui reaction curves from .csv exported from instrument.}
\usage{
imp_csv_sk_rc(x)
}
\arguments{
\item{x}{Path to the .csv file.}
}
\value{
A tibble of class "csv_sk_rc".
}
\description{
Import Sekisui reaction curves from .csv exported from instrument.
}
