% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_csv.R
\name{imp_csv_sk_res}
\alias{imp_csv_sk_res}
\title{Import Sekisui results from .csv exported from instrument.}
\usage{
imp_csv_sk_res(x)
}
\arguments{
\item{x}{Path to the .csv file.}
}
\value{
A tibble of class "csv_sk_res".
}
\description{
Import Sekisui results from .csv exported from instrument.
}
