% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVO.R
\name{imp_evo_spectra}
\alias{imp_evo_spectra}
\title{(WIP) Import absorption spectra results from EVO spectrophotometer}
\usage{
imp_evo_spectra(file_path)
}
\arguments{
\item{file_path}{}
}
\value{
a tibble
}
\description{
(WIP) Import absorption spectra results from EVO spectrophotometer
}
