% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kroma.R
\name{imp_kr_cal}
\alias{imp_kr_cal}
\title{Import calibration results from Kroma or Kroma Plus database.}
\usage{
imp_kr_cal(cal_ids, conn = connect_to_kr_dbi())
}
\arguments{
\item{cal_ids}{an integer vector of calibration ids}

\item{conn}{a connection object like the one returned from connect_to_kr_dbi}
}
\value{
a tibble of class "kr_cal"
}
\description{
Import calibration results from Kroma or Kroma Plus database.
}
