% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kroma.R
\name{imp_kr_results}
\alias{imp_kr_results}
\title{Import results from Kroma or Kroma Plus database.}
\usage{
imp_kr_results(conn = connect_to_kr_dbi())
}
\arguments{
\item{conn}{a connection object like the one returned from connect_to_kr_dbi}
}
\value{
a tibble of class "kr_results"
}
\description{
Import results from Kroma or Kroma Plus database.
}
