% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lida_300.R
\name{imp_l300_items}
\alias{imp_l300_items}
\title{Import the item parameters from Lida 300 database (WIP)}
\usage{
imp_l300_items(conn = connect_to_l300_dbi())
}
\arguments{
\item{conna}{connection object like the one returned from connect_to_l300_dbi}
}
\value{
a tibble of class "l300_items"
}
\description{
Import the item parameters from Lida 300 database (WIP)
}
