% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lida_500.R
\name{imp_l500_cal}
\alias{imp_l500_cal}
\title{Import calibration results from Lida 500 database.}
\usage{
imp_l500_cal(item_names, conn = connect_to_l500_dbi())
}
\arguments{
\item{item_names}{A character vector of ITEM_NAMEs as found in Sekisui database.}

\item{conn}{a connection object like the one returned from connect_to_sk_dbi}
}
\value{
a tibble of class "l500_cal"
}
\description{
Import calibration results from Lida 500 database.
}
