% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lida_500.R
\name{imp_l500_items}
\alias{imp_l500_items}
\title{Import the item parameters from Lida 500 database}
\usage{
imp_l500_items(conn = connect_to_l500_dbi())
}
\arguments{
\item{conn}{a connection object like the one returned from connect_to_l500_dbi}
}
\value{
a tibble of class "l500_items"
}
\description{
Import the item parameters from Lida 500 database
}
