% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lida_500.R
\name{imp_l500_results}
\alias{imp_l500_results}
\alias{imp_l500_worklisttest}
\alias{imp_l500_archive}
\title{Import results from Lida 500 database.}
\usage{
imp_l500_results(
  conn = connect_to_l500_dbi(),
  tables = c("all", "worklist", "archive")
)

imp_l500_worklisttest(conn = connect_to_l500_dbi())

imp_l500_archive(conn = connect_to_l500_dbi())
}
\arguments{
\item{conn}{a connection object like the one returned from connect_to_l500_dbi}

\item{tables}{the tables to import from}
}
\value{
imp_l500_worklisttest imports results currently stored in WorkListTest table.
imp_l500_archive imports archived results.
imp_l500_results can import from either or both tables.
}
\description{
Import results from Lida 500 database.
}
