% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sekisui.R
\name{imp_sk_caldata}
\alias{imp_sk_caldata}
\title{Import current calibration data from Sekisui for given methods.
`r lifecycle::badge('experimental')`}
\usage{
imp_sk_caldata(item_names, conn = connect_to_sk_dbi())
}
\arguments{
\item{item_names}{A character vector of ITEM_NAMEs as found in Sekisui database.}

\item{conn}{a connection object like the one returned from connect_to_sk_dbi}
}
\value{
a tibble of class "sk_cal"
}
\description{
Import current calibration data from Sekisui for given methods.
`r lifecycle::badge('experimental')`
}
