% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sekisui.R
\name{imp_sk_calrep}
\alias{imp_sk_calrep}
\title{Import the calibration replicates from Sekisui SK500's database.
`r lifecycle::badge('experimental')`}
\usage{
imp_sk_calrep(x, conn = connect_to_sk_dbi())
}
\arguments{
\item{x}{Object of class "sk_cal"}

\item{conn}{a connection object like the one returned from connect_to_sk_dbi}
}
\value{
a tibble of class "sk_cal_rep"
}
\description{
Import the calibration replicates from Sekisui SK500's database.
`r lifecycle::badge('experimental')`
}
