% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sekisui.R
\name{imp_sk_items}
\alias{imp_sk_items}
\title{Import the item parameters from Sekisui database}
\usage{
imp_sk_items(conn = connect_to_sk_dbi())
}
\arguments{
\item{conna}{connection object like the one returned from connect_to_sk_dbi}
}
\value{
a tibble of class "sk_items"
}
\description{
Import the item parameters from Sekisui database
}
