% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sekisui.R
\name{imp_sk_rc}
\alias{imp_sk_rc}
\alias{imp_sk_rc.numeric}
\alias{imp_sk_rc.sk_results}
\title{Import reaction curves from Sekisui SK500
`r lifecycle::badge('experimental')`}
\usage{
imp_sk_rc(x, conn = connect_to_sk_dbi(), ...)

\method{imp_sk_rc}{numeric}(x, conn = connect_to_sk_dbi())

\method{imp_sk_rc}{sk_results}(x, conn = connect_to_sk_dbi())
}
\arguments{
\item{x}{Describir...}
}
\value{
a tibble of class "sk_rc".
}
\description{
Import reaction curves from Sekisui SK500
`r lifecycle::badge('experimental')`
}
\section{Methods (by class)}{
\itemize{
\item \code{imp_sk_rc(numeric)}: method

\item \code{imp_sk_rc(sk_results)}: method

}}
