% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importr2-package.R
\docType{package}
\name{importr2-package}
\alias{importr2}
\alias{importr2-package}
\title{importr2: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
