% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{l500_con_strings}
\alias{l500_con_strings}
\title{Build connection string for Lida 500 database}
\usage{
l500_con_strings(db_file_path)
}
\arguments{
\item{db_file_path}{path to database file (.mdb)}
}
\value{
A connection string for Lida 500 database.
}
\description{
Build connection string for Lida 500 database
}
