% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_cal.R
\name{wrangle_cal}
\alias{wrangle_cal}
\alias{wrangle_cal.l500_cal}
\alias{wrangle_cal.sk_cal}
\alias{wrangle_cal.kr_cal}
\alias{wrangle_cal.l300_cal}
\title{Wrangle calibration results imported from different instrument databases into a common data structure}
\usage{
wrangle_cal(x, ...)

\method{wrangle_cal}{l500_cal}(x)

\method{wrangle_cal}{sk_cal}(x)

\method{wrangle_cal}{kr_cal}(x, instrument = c("Kroma", "Kroma Plus"))

\method{wrangle_cal}{l300_cal}(x)
}
\arguments{
\item{x}{data imported from an instrument database}

\item{...}{additional arguments used by methods}
}
\value{
a tibble with columns 'method', 'value', and 'od'.
}
\description{
Wrangle calibration results imported from different instrument databases into a common data structure
}
\section{Methods (by class)}{
\itemize{
\item \code{wrangle_cal(l500_cal)}: method

\item \code{wrangle_cal(sk_cal)}: method

\item \code{wrangle_cal(kr_cal)}: method

\item \code{wrangle_cal(l300_cal)}: method

}}
