% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_rc.R
\name{wrangle_rc}
\alias{wrangle_rc}
\alias{wrangle_rc.sk_rc}
\title{Wrangle reaction curves imported from different instrument databases into a common data structure}
\usage{
wrangle_rc(x, ...)

\method{wrangle_rc}{sk_rc}(x, ...)
}
\arguments{
\item{x}{data imported from an instrument database}

\item{...}{additional arguments used by methods}
}
\value{
a tibble with columns 'instrument','id_rc', 'cycle', 'abs1' and 'abs2'.
}
\description{
Wrangle reaction curves imported from different instrument databases into a common data structure
}
\section{Methods (by class)}{
\itemize{
\item \code{wrangle_rc(sk_rc)}: method

}}
