% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_results.R
\name{wrangle_results}
\alias{wrangle_results}
\alias{wrangle_results.l500_results}
\alias{wrangle_results.sk_results}
\alias{wrangle_results.kr_results}
\alias{wrangle_results.csv_kr_res}
\title{Wrangle results imported from different instrument databases into a common data structure}
\usage{
wrangle_results(x, ...)

\method{wrangle_results}{l500_results}(x)

\method{wrangle_results}{sk_results}(x)

\method{wrangle_results}{kr_results}(x, instrument = c("Kroma", "Kroma Plus"))

\method{wrangle_results}{csv_kr_res}(x, instrument = c("Kroma", "Kroma Plus"))
}
\arguments{
\item{x}{data imported from an instrument database}

\item{...}{additional arguments used by methods}

\item{instrument}{either "Kroma" or "Kroma Plus"}
}
\value{
a tibble with columns 'date', 'method', 'instrument', 'sample' and 'result'.
}
\description{
Wrangle results imported from different instrument databases into a common data structure
}
\section{Methods (by class)}{
\itemize{
\item \code{wrangle_results(l500_results)}: method

\item \code{wrangle_results(sk_results)}: method

\item \code{wrangle_results(kr_results)}: method

\item \code{wrangle_results(csv_kr_res)}: method

}}
