% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sekisui.R
\name{fct_meanSek}
\alias{fct_meanSek}
\title{Compute a mean like sekisui SK500 (drop one, mean of the rest)
`r lifecycle::badge('stable')`}
\usage{
fct_meanSek(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
A single number.
}
\description{
Compute a mean like sekisui SK500 (drop one, mean of the rest)
`r lifecycle::badge('stable')`
}
