% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sekisui.R
\name{imp_sk_rc}
\alias{imp_sk_rc}
\alias{imp_sk_rc.integer}
\alias{imp_sk_rc.sk_results}
\title{Import reaction curves from Sekisui SK500
`r lifecycle::badge('experimental')`}
\usage{
imp_sk_rc(x, ...)

\method{imp_sk_rc}{integer}(x)

\method{imp_sk_rc}{sk_results}(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
A single number.
}
\description{
Import reaction curves from Sekisui SK500
`r lifecycle::badge('experimental')`
}
\section{Methods (by class)}{
\itemize{
\item \code{imp_sk_rc(integer)}: method

\item \code{imp_sk_rc(sk_results)}: method

}}
